# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Diet Assistant API
  description: |
    RESTful API for AI-powered diet planning and nutrition management system.
    
    ## Features
    - JWT-based authentication
    - Food inventory management
    - Meal tracking and recording
    - AI-powered meal plan generation
    - Nutrition analysis and statistics
    - Conversation with AI assistant
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication.
    Include the token in the Authorization header: `Bearer {token}`
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9090/api/v1
    description: Development server
  - url: https://api.yourdomain.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
          example: 40001
        message:
          type: string
          example: "Invalid parameters"
        error:
          type: string
          example: "Detailed error message"
        timestamp:
          type: integer
          format: int64
          example: 1699999999
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: "success"
        timestamp:
          type: integer
          format: int64
          example: 1699999999
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "user"
        password:
          type: string
          format: password
          example: "password123"
    TokenResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                access_token:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                refresh_token:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                expires_in:
                  type: integer
                  example: 900
                user:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    username:
                      type: string
                      example: "user"
    NutritionData:
      type: object
      properties:
        protein:
          type: number
          format: float
          example: 31.0
        carbs:
          type: number
          format: float
          example: 0.0
        fat:
          type: number
          format: float
          example: 3.6
        fiber:
          type: number
          format: float
          example: 0.0
        calories:
          type: number
          format: float
          example: 165.0
...
    MealFood:
      type: object
      properties:
        food_id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Chicken Breast"
        amount:
          type: number
          format: float
          example: 200
        unit:
          type: string
          example: "g"
    Meal:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        user_id:
          type: integer
          format: int64
          example: 1
        meal_date:
          type: string
          format: date
          example: "2024-01-01"
        meal_type:
          type: string
          enum: [breakfast, lunch, dinner, snack]
          example: "breakfast"
        foods:
          type: array
          items:
            $ref: '#/components/schemas/MealFood'
        nutrition:
          $ref: '#/components/schemas/NutritionData'
        notes:
          type: string
          example: "Healthy breakfast"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Plan:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        user_id:
          type: integer
          format: int64
          example: 1
        plan_date:
          type: string
          format: date
          example: "2024-01-02"
        meal_type:
          type: string
          enum: [breakfast, lunch, dinner, snack]
          example: "breakfast"
        foods:
          type: array
          items:
            $ref: '#/components/schemas/MealFood'
        nutrition:
          $ref: '#/components/schemas/NutritionData'
        status:
          type: string
          enum: [pending, completed, skipped]
          example: "pending"
        ai_reasoning:
          type: string
          example: "High protein breakfast to start your day"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive access tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate current tokens
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /ai/chat:
    post:
      tags:
        - AI
      summary: Chat with AI
      description: Have a conversation with AI assistant
      operationId: chatWithAI
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  example: "What should I eat for dinner?"
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          response:
                            type: string
                          tokens_used:
                            type: integer
...
  /dashboard:
    get:
      tags:
        - Dashboard
      summary: Get dashboard data
      description: Get comprehensive dashboard data
      operationId: getDashboard
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          recent_meals:
                            type: array
                            items:
                              $ref: '#/components/schemas/Meal'
                          upcoming_plans:
                            type: array
                            items:
                              $ref: '#/components/schemas/Plan'